Rename SecureTokenTransferProtocol to TokenHolder
