+++ docs/RMRK/extension/tokenHolder/RMRKTokenHolder.md
@@ -1,19 +1,19 @@
-# RMRKSecureTokenTransferProtocol
+# RMRKTokenHolder
 
+*RMRK team*
 
+> RMRKTokenHolder
 
+Smart contract of a token holder RMRK extension.
 
-
-
-
-
+*The RMRKTokenHolder extension is capable of holding ERC-20, ERC-721, and ERC-1155 tokens.*
 
 ## Methods
 
 ### balanceOfToken
 
 ```solidity
-function balanceOfToken(address tokenContract, enum IRMRKSecureTokenTransferProtocol.TokenType tokenType, uint256 tokenId, uint256 heldTokenId) external view returns (uint256)
+function balanceOfToken(address tokenContract, enum IRMRKTokenHolder.TokenType tokenType, uint256 tokenId, uint256 heldTokenId) external view returns (uint256)
 ```
 
 Used to retrieve the given token&#39;s balance of given token
@@ -25,7 +25,7 @@ Used to retrieve the given token&#39;s balance of given token
 | Name | Type | Description |
 |---|---|---|
 | tokenContract | address | The address of the held token&#39;s smart contract |
-| tokenType | enum IRMRKSecureTokenTransferProtocol.TokenType | The type of the token being checked for balance |
+| tokenType | enum IRMRKTokenHolder.TokenType | The type of the token being checked for balance |
 | tokenId | uint256 | The ID of the token being checked for balance |
 | heldTokenId | uint256 | The ID of the held token of which the balance is being retrieved |
 
@@ -111,7 +111,7 @@ function supportsInterface(bytes4 interfaceId) external view returns (bool)
 ### transferHeldTokenFromToken
 
 ```solidity
-function transferHeldTokenFromToken(address tokenContract, enum IRMRKSecureTokenTransferProtocol.TokenType tokenType, uint256 tokenId, uint256 heldTokenId, uint256 amount, address to, bytes data) external nonpayable
+function transferHeldTokenFromToken(address tokenContract, enum IRMRKTokenHolder.TokenType tokenType, uint256 tokenId, uint256 heldTokenId, uint256 amount, address to, bytes data) external nonpayable
 ```
 
 Transfer held tokens from a specific token.
@@ -123,7 +123,7 @@ Transfer held tokens from a specific token.
 | Name | Type | Description |
 |---|---|---|
 | tokenContract | address | The address of the held token&#39;s smart contract |
-| tokenType | enum IRMRKSecureTokenTransferProtocol.TokenType | The type of the token being transferred |
+| tokenType | enum IRMRKTokenHolder.TokenType | The type of the token being transferred |
 | tokenId | uint256 | The ID of the token to transfer the held token from |
 | heldTokenId | uint256 | The ID of the held token to transfer |
 | amount | uint256 | The number of held tokens to transfer |
@@ -133,7 +133,7 @@ Transfer held tokens from a specific token.
 ### transferHeldTokenToToken
 
 ```solidity
-function transferHeldTokenToToken(address tokenContract, enum IRMRKSecureTokenTransferProtocol.TokenType tokenType, uint256 tokenId, uint256 heldTokenId, uint256 amount, bytes data) external nonpayable
+function transferHeldTokenToToken(address tokenContract, enum IRMRKTokenHolder.TokenType tokenType, uint256 tokenId, uint256 heldTokenId, uint256 amount, bytes data) external nonpayable
 ```
 
 Transfer tokens to a specific holder token.
@@ -145,7 +145,7 @@ Transfer tokens to a specific holder token.
 | Name | Type | Description |
 |---|---|---|
 | tokenContract | address | The address of the token smart contract |
-| tokenType | enum IRMRKSecureTokenTransferProtocol.TokenType | The type of the token being transferred |
+| tokenType | enum IRMRKTokenHolder.TokenType | The type of the token being transferred |
 | tokenId | uint256 | The ID of the token to transfer the tokens to |
 | heldTokenId | uint256 | The ID of the held token to transfer |
 | amount | uint256 | The number of ERC-20 tokens to transfer |
@@ -158,7 +158,7 @@ Transfer tokens to a specific holder token.
 ### ReceivedToken
 
 ```solidity
-event ReceivedToken(address indexed tokenContract, enum IRMRKSecureTokenTransferProtocol.TokenType tokenType, uint256 indexed toTokenId, uint256 heldTokenId, address indexed from, uint256 amount)
+event ReceivedToken(address indexed tokenContract, enum IRMRKTokenHolder.TokenType tokenType, uint256 indexed toTokenId, uint256 heldTokenId, address indexed from, uint256 amount)
 ```
 
 Used to notify listeners that the token received held tokens.
@@ -170,7 +170,7 @@ Used to notify listeners that the token received held tokens.
 | Name | Type | Description |
 |---|---|---|
 | tokenContract `indexed` | address | The address of the held token&#39;s smart contract |
-| tokenType  | enum IRMRKSecureTokenTransferProtocol.TokenType | The type of the held token being received |
+| tokenType  | enum IRMRKTokenHolder.TokenType | The type of the held token being received |
 | toTokenId `indexed` | uint256 | The ID of the token receiving the held tokens |
 | heldTokenId  | uint256 | The ID of the held token being received |
 | from `indexed` | address | The address of the account from which the tokens are being transferred |
@@ -179,7 +179,7 @@ Used to notify listeners that the token received held tokens.
 ### TransferredToken
 
 ```solidity
-event TransferredToken(address indexed tokenContract, enum IRMRKSecureTokenTransferProtocol.TokenType tokenType, uint256 indexed fromTokenId, uint256 heldTokenId, address indexed to, uint256 amount)
+event TransferredToken(address indexed tokenContract, enum IRMRKTokenHolder.TokenType tokenType, uint256 indexed fromTokenId, uint256 heldTokenId, address indexed to, uint256 amount)
 ```
 
 Used to notify the listeners that the ERC-20 tokens have been transferred.
@@ -191,7 +191,7 @@ Used to notify the listeners that the ERC-20 tokens have been transferred.
 | Name | Type | Description |
 |---|---|---|
 | tokenContract `indexed` | address | The address of the smart contract of the token being transferred |
-| tokenType  | enum IRMRKSecureTokenTransferProtocol.TokenType | The type of the token being transferred |
+| tokenType  | enum IRMRKTokenHolder.TokenType | The type of the token being transferred |
 | fromTokenId `indexed` | uint256 | The ID of the token from which the held tokens have been transferred |
 | heldTokenId  | uint256 | The Id of the held token being transferred |
 | to `indexed` | address | The address receiving the ERC-20 tokens |
