--- docs/console.md
@@ -1,12 +1,0 @@
-# console
-
-
-
-
-
-
-
-
-
-
-
