+++ test/interfaces.ts
@@ -1,5 +1,5 @@
 const IERC165 = '0x01ffc9a7';
-const IRMRKSecureTokenTransferProtocol = '0x99782111';
+const IRMRKTokenHolder = '0x99782111';
 const IERC721 = '0x80ac58cd';
 const IERC721Metadata = '0x5b5e139f';
 const IOtherInterface = '0xffffffff';
@@ -22,7 +22,7 @@ const IRMRKImplementation = '0x524D524B'; // "RMRK" in ASCII hex
 
 export {
   IERC165,
-  IRMRKSecureTokenTransferProtocol,
+  IRMRKTokenHolder,
   IERC721,
   IERC721Metadata,
   IOtherInterface,
